tests: move trust test to proper package
